+++ io_chunks/exceptions.py
@@ -1,0 +1,3 @@
+class ClosedStreamError(ValueError):
+    def __init__(self):
+        super().__init__("I/O operation on closed chunk")