+++ CHANGELOG.md
@@ -15,11 +15,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add tests to ensure that `RawIOChunk` behaves consistently with all the other
   Now `RawIOChunk` implements `IO` interface.
 - Implement `IO` interface for `RawIOChunk`.
+- Add `exceptions.py`.
 ### Removed
 - Remove Python 2 support.
 ### Changed
 - Change `README.rst` to `README.md`.
 - Replace `setup.py` with `pyproject.toml`.
+- Renamed `chunks.py` to `raw_io_chunk.py`.
 ### Fixed
 - Now returns an empty bytes instead of raising `EOFError` when the underlying
   stream is empty.